+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #130 Allow to deactivate/activate Doctors
 
 **Changed**
 
@@ -16,7 +17,8 @@ Changelog
 **Fixed**
 
 - #133 System clearing patient first name on editing patient
-
+- #132 Autofill Batch field in Sample Add form when Batch defined in request
+- #131 Autofill ClientPatientID field in Sample Add form when Patient defined
 
 **Security**